Crate actix_multipart

Multipart form support for Actix Web.

Examples

use actix_web::{post, App, HttpServer, Responder};

use actix_multipart::form::{json::Json as MPJson, tempfile::TempFile, MultipartForm};
use serde::Deserialize;

#[derive(Debug, Deserialize)]
struct Metadata {
    name: String,
}

#[derive(Debug, MultipartForm)]
struct UploadForm {
    #[multipart(limit = "100MB")]
    file: TempFile,
    json: MPJson<Metadata>,
}

#[post("/videos")]
pub async fn post_video(MultipartForm(form): MultipartForm<UploadForm>) -> impl Responder {
    format!(
        "Uploaded file {}, with size: {}",
        form.json.name, form.file.size
    )
}

#[actix_web::main]
async fn main() -> std::io::Result<()> {
    HttpServer::new(move || App::new().service(post_video))
        .bind(("127.0.0.1", 8080))?
        .run()
        .await
}

Re-exports

pub use self::test::create_form_data_payload_and_headers;

pub use self::test::create_form_data_payload_and_headers_with_boundary;

Modules

form

Process and extract typed data from a multipart stream.

test

Structs

Field

A single field in a multipart stream

Multipart

The server-side implementation of multipart/form-data requests.

Enums

MultipartError

A set of errors that can occur during parsing multipart streams.