Crate actix_multipart

Multipart form support for Actix Web.

Examples

use actix_web::{post, App, HttpServer, Responder};

use actix_multipart::form::{json::Json as MpJson, tempfile::TempFile, MultipartForm};
use serde::Deserialize;

#[derive(Debug, Deserialize)]
struct Metadata {
    name: String,
}

#[derive(Debug, MultipartForm)]
struct UploadForm {
    #[multipart(limit = "100MB")]
    file: TempFile,
    json: MpJson<Metadata>,
}

#[post("/videos")]
pub async fn post_video(MultipartForm(form): MultipartForm<UploadForm>) -> impl Responder {
    format!(
        "Uploaded file {}, with size: {}",
        form.json.name, form.file.size
    )
}

#[actix_web::main]
async fn main() -> std::io::Result<()> {
    HttpServer::new(move || App::new().service(post_video))
        .bind(("127.0.0.1", 8080))?
        .run()
        .await
}

cURL request:

curl -v --request POST \
  --url http://localhost:8080/videos \
  -F 'json={"name": "Cargo.lock"};type=application/json' \
  -F file=@./Cargo.lock

Modules

form

Process and extract typed data from a multipart stream.

test

Multipart testing utilities.

Structs

Field

A single field in a multipart stream.

LimitExceeded

Error type returned from Field::bytes() when field data is larger than limit.

Multipart

The server-side implementation of multipart/form-data requests.

Enums

MultipartError

A set of errors that can occur during parsing multipart streams.