actix_multipart
Module form

Process and extract typed data from a multipart stream.

Modules

bytes

Reads a field into memory.

json

Deserializes a field as JSON.

tempfile

Writes a field to a temporary file on disk.

text

Deserializes a field from plain text.

Structs

Limits

Used to keep track of the remaining limits for the form and current field.

MultipartForm

Typed multipart/form-data extractor.

MultipartFormConfig

MultipartForm extractor configuration.

Traits

FieldReader

Trait that data types to be used in a multipart form struct should implement.

MultipartCollect

Trait that allows a type to be used in the MultipartForm extractor.

Derive Macros

MultipartForm

Implements MultipartCollect for a struct so that it can be used with the MultipartForm extractor.